refactor: move jpeg compression (#751)
* refactor: move jpeg compression

* refactor: bump version
